% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS.R
\name{CS0}
\alias{CS0}
\title{Internal: Continuous Spike}
\usage{
CS0(
  Betah1,
  Betah2,
  Sigmah1,
  Sigmah2,
  kappa0,
  tau20,
  zeta10,
  zeta20,
  mg,
  niter = 1000,
  burnin = 500,
  nthin = 2,
  a1 = 0.1,
  a2 = 1,
  c1 = 0.1,
  c2 = 1,
  sigma2 = 10^-3,
  snpnames,
  genename
)
}
\arguments{
\item{Betah1}{A numerical vector of length mg representing the regression coefficients for the first trait.}

\item{Betah2}{A numerical vector of length mg representing the regression coefficients for the second trait.}

\item{Sigmah1}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix of Betah1.}

\item{Sigmah2}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix of Betah2.}

\item{kappa0}{Initial value for kappa.}

\item{tau20}{Initial value for tau2.}

\item{zeta10}{Initial value for zeta1.}

\item{zeta20}{Initial value for zeta2.}

\item{mg}{Number of variables in the group.}

\item{niter}{Number of iterations for the Gibbs sampler.}

\item{burnin}{Number of burn-in iterations.}

\item{nthin}{The lag of the iterations used for the posterior analysis is defined (or thinning rate).}

\item{a1, a2}{Hyperparameters of kappa. Default is a1=0.1 and a2=1.}

\item{c1, c2}{Hyperparameters of tau2. Default is c1=0.1 and c2=1.}

\item{sigma2}{Variance of spike (multivariate normal distribution with a diagonal covariance matrix with small variance) representing the null effect distribution. Default is 10^-3.}

\item{snpnames}{Names of variables for the group.}

\item{genename}{Name of group.}
}
\description{
Internal: Continuous Spike
}
