% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAJC1.R
\docType{data}
\name{DNAJC1}
\alias{DNAJC1}
\title{Gene DNAJC1 from BCAC and Epithyr  studies}
\format{
A list which contains two matrices for the summary statistics of each study.
\describe{
  \item{Breast}{Summary statitics of breast cancer including the name of SNPs, beta and se}
  \item{Thyroid}{Summary statitics of thyroid cancer including the name of SNPs, beta and se}
}
}
\usage{
DNAJC1
}
\description{
The summary statistics data for DNAJC1 protein coding gene including beta and standard
error for pleiotropy investigation of breast and thyroid cancers.
The summary statistics of breast and thyroid  cancer are extracted from BCAC and Epithyr  (Baghgalaki et al., 2021b) studies, respectively.
}
\references{
\enumerate{
\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\seealso{
\code{\link{GCPBayes}}
}
\keyword{datasets}
