% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS.R
\name{e2_Monte_Carlo_EM}
\alias{e2_Monte_Carlo_EM}
\title{Internal: e2_Monte_Carlo_EM}
\usage{
e2_Monte_Carlo_EM(
  Betah,
  Sigmah,
  kappa0 = kappa0,
  kappastar0 = kappastar0,
  sigma20 = sigma20,
  s20 = s20,
  m,
  K,
  a1 = a1,
  a2 = a2,
  d1 = d1,
  d2 = d2,
  c1 = c1,
  c2 = c2,
  e2 = e2,
  snpnames,
  genename
)
}
\arguments{
\item{Betah}{A matrix of dimension K*m represents the regression coefficients. Each row of this matrix includes the regression coefficients for each trait.}

\item{Sigmah}{A symmetric block-diagonal matrix of dimension K*m is used. Each block of this matrix shows a positive definite covariance matrix which is an estimated covariance matrix of each trait.}

\item{kappa0}{Initial value for kappa.}

\item{kappastar0}{Initial value for kappastar.}

\item{sigma20}{Initial value for sigma2.}

\item{s20}{Initial value for s2.}

\item{m}{Number of variables in the group.}

\item{K}{Number of traits.}

\item{a1, a2}{Hyperparameters of kappa. Default is a1=0.1 and a2=0.1.}

\item{d1, d2}{Hyperparameters of sigma2. Default is d1=0.1 and d2=0.1.}

\item{c1, c2}{Hyperparameters of kappastar. Default is c1=0.1 and c2=0.1.}

\item{e2}{Initial value for doing Monte Carlo EM algorithm to estimate hyperparameter of s2.}

\item{snpnames}{Names of variables for the group.}

\item{genename}{Name of group.}
}
\description{
Internal: e2_Monte_Carlo_EM
}
