% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCPBayes.R
\docType{package}
\name{GCPBayes}
\alias{GCPBayes}
\title{GCPBayes Package}
\description{
Run a Gibbs sampler for a multivariate Bayesian sparse group selection model with Dirac, continuous and hierarchical spike prior for detecting pleiotropic effects on multiple traits. This package is designed for summary statistics containing estimated regression coefficients and their estimated covariance matrices.
}
\references{
\enumerate{
\item
Baghfalaki, T., Sugier, P. E., Truong, T., Pettitt, A. N., Mengersen, K., & Liquet, B. (2021). Bayesian meta analysis models for cross cancer genomic investigation of pleiotropic effects using group structure. \emph{Statistics in Medicine}, \strong{40}(6), 1498-1518.

\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, \email{t.baghfalaki@modares.ac.ir}
}
\keyword{CS}
\keyword{DS}
\keyword{HS}
\keyword{MCMCplot}
\keyword{summaryCS}
\keyword{summaryDS}
\keyword{summaryHS}
