% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARP2.R
\docType{data}
\name{PARP2}
\alias{PARP2}
\title{Gene PARP2 from CECILE study}
\format{
A list which contains two matrices.
\describe{
  \item{Breast}{Summary statitics of breast cancer including the name of SNPs, beta and se}
  \item{Thyroid}{Summary statitics of thyroid cancer including the name of SNPs, beta and se}
}
}
\usage{
PARP2
}
\description{
A list containing the individual level data for gene PARP2 including genotypes and phenotypes for pleiotropy
investigation of breast and thyroid cancers. It is from
CECILE study, a French population-based case-control study on breast cancer and from the French
studies included in the EPITHYR consortium on thyroid cancer.
}
\references{
\enumerate{
\item
Baghfalaki, T., Sugier, P. E., Truong, T., Pettitt, A. N., Mengersen, K., & Liquet, B. (2021). Bayesian meta analysis models for cross cancer genomic investigation of pleiotropic effects using group structure. \emph{Statistics in Medicine}, \strong{40}(6), 1498-1518.

\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\seealso{
\code{\link{GCPBayes}}
}
\keyword{datasets}
