% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARP2_summary.R
\docType{data}
\name{PARP2_summary}
\alias{PARP2_summary}
\title{Summary statistics of gene PARP2 from CECILE study}
\format{
A list which contains two matrices for the summary statistics of each study.
\describe{
  \item{Breast}{Summary statitics of breast cancer including the name of SNPs, beta and se}
  \item{Thyroid}{Summary statitics of thyroid cancer including the name of SNPs, beta and se}
}
}
\usage{
PARP2_summary
}
\description{
The summary statistics data for PARP2 protein coding gene including beta and standard error for pleiotropy investigation of breast and thyroid cancers.
}
\references{
\enumerate{
\item
Baghfalaki, T., Sugier, P. E., Truong, T., Pettitt, A. N., Mengersen, K., & Liquet, B. (2021). Bayesian meta analysis models for cross cancer genomic investigation of pleiotropic effects using group structure. \emph{Statistics in Medicine}, \strong{40}(6), 1498-1518.

\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\seealso{
\code{\link{GCPBayes}}
}
\keyword{datasets}
