% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DS.R
\name{plot_acf_DS}
\alias{plot_acf_DS}
\title{ACF plot for the output of DS}
\usage{
plot_acf_DS(
  DS,
  kappa = FALSE,
  sigma2 = FALSE,
  Beta1 = NULL,
  Beta2 = NULL,
  nrow1 = 1,
  nrow2 = 1
)
}
\arguments{
\item{DS}{All the generated results by DS function.}

\item{kappa}{logical; if TRUE, the ACF plot of kappa is drawan.}

\item{sigma2}{logical; if TRUE, the ACF plot of sigma2 is drawan.}

\item{Beta1}{A vector of indexes for drawing ACF plot of Beta1. If NULL, the ACF plot of Beta1 is not drawn.}

\item{Beta2}{A vector of indexes for drawing ACF plot of Beta2. If NULL, the ACF plot of Beta2 is not drawn.}

\item{nrow1}{Number of rows of the ACF plot of Beta1.}

\item{nrow2}{Number of rows of the ACF plot of Beta2.}
}
\description{
ACF plot for the output of DS
}
\seealso{
  \code{\link{DS}}
}
\examples{
mg=10
SD=diag(0.05,mg)
corr=0.5
R=matrix(corr,mg,mg)+(1-corr)*diag(mg)
Sigma=crossprod(crossprod(SD,R),SD)
sign=rbinom(mg,1,.5)
sign[sign==0]=-1
Betat=rep(1,mg)
Betat=Betat*sign
Betah=mvtnorm::rmvnorm(2,Betat,Sigma)

snpnames=1:mg
genename="simulated_data"


Betah1=Betah[1,]; Betah2=Betah[2,];
Sigmah1=Sigma; Sigmah2=Sigma;


niter=5000;burnin=1000

res1 = DS(Betah1, Betah2,
          Sigmah1, Sigmah2,
          kappa0=0.5, sigma20=1,
          mg=mg, niter=niter, burnin=burnin,
          nchains=1, nthin=2, a1=0.1, a2=0.1, d1=0.1, d2=0.1,
          snpnames, genename)

res2 = DS(Betah1, Betah2,
          Sigmah1, Sigmah2,
          kappa0=c(0.5,.25,0.6), sigma20=c(1,1.2,1.5),
          mg=mg, niter=niter, burnin=burnin,
          nchains=3, nthin=2, a1=0.1, a2=0.1, d1=0.1, d2=0.1,
          snpnames, genename)


plot_acf_DS(res1, kappa=TRUE, sigma2=TRUE,
            Beta1=1:mg, Beta2=1:mg)


plot_acf_DS(res2, kappa=TRUE, sigma2=TRUE,
            Beta1=1:mg, Beta2=1:mg)




}
