% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HS.R
\name{plot_trace_HS}
\alias{plot_trace_HS}
\title{Trace plot for the output of HS}
\usage{
plot_trace_HS(
  HS,
  kappa = FALSE,
  kappastar = FALSE,
  sigma2 = FALSE,
  Beta1 = NULL,
  Beta2 = NULL,
  nrow1 = 1,
  nrow2 = 1
)
}
\arguments{
\item{HS}{All the generated results by HS function.}

\item{kappa}{logical; if TRUE, the trace plot of kappa is drawan.}

\item{kappastar}{logical; if TRUE, the density plot of kappastar is drawan.}

\item{sigma2}{logical; if TRUE, the trace plot of sigma2 is drawan.}

\item{Beta1}{A vector of indexes for drawing trace plot of Beta1. If NULL, the trace plot of Beta1 is not drawn.}

\item{Beta2}{A vector of indexes for drawing trace plot of Beta2. If NULL, the trace plot of Beta2 is not drawn.}

\item{nrow1}{Number of rows of the trace plot of Beta1.}

\item{nrow2}{Number of rows of the trace plot of Beta2.}
}
\description{
Trace plot for the output of HS
}
\seealso{
  \code{\link{HS}}
}

\examples{
mg=10
SD=diag(0.05,mg)
corr=0.5
R=matrix(corr,mg,mg)+(1-corr)*diag(mg)
Sigma=crossprod(crossprod(SD,R),SD)
sign=rbinom(mg,1,.5)
sign[sign==0]=-1
Betat=rep(1,mg)
Betat=Betat*sign
Betah=mvtnorm::rmvnorm(2,Betat,Sigma)

snpnames=1:mg
genename="simulated_data"


#######################
Betah1=Betah[1,]; Betah2=Betah[2,];
Sigmah1=Sigma; Sigmah2=Sigma;
a1=0.1; a2=0.1; d1=0.1; d2=0.1; c1=1; c2=1; e2=1; N00=100



res1 = HS(Betah1, Betah2,
          Sigmah1, Sigmah2, kappa0 = 0.5, kappastar0=0.5, sigma20 =1, s20 =1,
          mg=mg, niter=1000, burnin=500, nthin=3, nchains=1, a1=a1, a2=a2, d1=d1,
          d2=d2, c1=c1, c2=c2, e2=e2, snpnames, genename)



res2 = HS(Betah1, Betah2,
          Sigmah1, Sigmah2, kappa0 = c(0.5,0.25,0.7), kappastar0=c(0.5,0.25,0.2),
          sigma20 =c(1,1.5,2), s20 =c(1,1.5,2),
          mg=mg, niter=1000, burnin=500, nthin=3, nchains=3, a1=a1, a2=a2, d1=d1,
          d2=d2, c1=c1, c2=c2, e2=e2, snpnames, genename)



plot_trace_HS(res1, kappa=TRUE, kappastar=TRUE,  sigma2=TRUE,
              Beta1=1:mg, Beta2=1:mg,
              nrow1 = 2, nrow2 = 2)

plot_trace_HS(res2, kappa=TRUE, kappastar=TRUE,  sigma2=TRUE,
              Beta1=1:mg, Beta2=1:mg,
              nrow1 = 2, nrow2 = 2)



























}
