% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CS.R
\name{reshape_CS}
\alias{reshape_CS}
\title{Internal : Reshape CS mcmc chain}
\usage{
reshape_CS(CS)
}
\arguments{
\item{CS}{All the generated results by CS function.}
}
\description{
Internal : Reshape CS mcmc chain
}
