% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DS.R
\name{reshape_DS}
\alias{reshape_DS}
\title{Internal : Reshape DS mcmc chain}
\usage{
reshape_DS(DS)
}
\arguments{
\item{DS}{All the generated results by DS function.}
}
\description{
Internal : Reshape DS mcmc chain
}
