% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HS.R
\name{reshape_HS}
\alias{reshape_HS}
\title{Internal : Reshape HS mcmc chain}
\usage{
reshape_HS(HS)
}
\arguments{
\item{HS}{All the generated results by HS function.}
}
\description{
Internal : Reshape HS mcmc chain
}
