% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulated_individual.R
\docType{data}
\name{Simulated_individual}
\alias{Simulated_individual}
\title{Simulated individual level data}
\format{
A list which contains the name of gene, the name of the SNPs, the vectors and the matrices for the summary statistics of each study.
\describe{
  \item{Study1}{The inividual level data including genotypes and phenotypes of Study 1}
  \item{Study2}{The inividual level data including genotypes and phenotypes of Study 2}
  \item{Study3}{The inividual level data including genotypes and phenotypes of Study 3}
}
}
\usage{
Simulated_individual
}
\description{
A list containing the individual level data including genotypes and phenotypes for pleiotropy investigation of three studies.
}
\references{
\enumerate{'
\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\seealso{
\code{\link{GCPBayes}}
}
\keyword{datasets}
