% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulated_individual_survival.R
\docType{data}
\name{Simulated_individual_survival}
\alias{Simulated_individual_survival}
\title{Simulated individual level survival data}
\format{
A list which contains the name of gene, the name of the SNPs, the vectors and the matrices for the summary statistics of each study.
\describe{
  \item{Study1}{The inividual level data including survival outcomes and gene expression data of Study 1}
  \item{Study2}{The inividual level data including survival outcomes and gene expression data of Study 2}
}
}
\usage{
Simulated_individual_survival
}
\description{
A list containing the individual level data including gene expression data for survival outcomes for pleiotropy investigation of two studies.
#' @name Simulated_individual_survival
}
\references{
\enumerate{'
\item
Baghfalaki, T., Sugier, Y. Asgari, P. E., Truong, & Liquet, B. (2021). GCPBayes:  An R Package for Studying Cross-Phenotype Genetic Associations with Group-Level Bayesian Meta-Analysis. \emph{Submitted}.
}
}
\seealso{
\code{\link{GCPBayes}}
}
\keyword{datasets}
