% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS.R
\name{CS}
\alias{CS}
\title{Continuous Spike}
\usage{
CS(
  Betah1,
  Betah2,
  Sigmah1,
  Sigmah2,
  kappa0,
  tau20,
  zeta10,
  zeta20,
  mg,
  niter = 1000,
  burnin = 500,
  nthin = 2,
  nchains = 1,
  a1 = 0.1,
  a2 = 1,
  c1 = 0.1,
  c2 = 1,
  sigma2 = 10^-3,
  snpnames,
  genename
)
}
\arguments{
\item{Betah1}{A numerical vector of length mg representing the regression coefficients for the first trait.}

\item{Betah2}{A numerical vector of length mg representing the regression coefficients for the second trait.}

\item{Sigmah1}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix of Betah1.}

\item{Sigmah2}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix of Betah2.}

\item{kappa0}{Initial value for kappa (its dimension is equal to nchains).}

\item{tau20}{Initial value for tau2 (its dimension is equal to nchains).}

\item{zeta10}{Initial value for zeta1 (its dimension is equal to nchains).}

\item{zeta20}{Initial value for zeta2 (its dimension is equal to nchains).}

\item{mg}{Number of variables in the group.}

\item{niter}{Number of iterations for the Gibbs sampler.}

\item{burnin}{Number of burn-in iterations.}

\item{nthin}{The lag of the iterations used for the posterior analysis is defined (or thinning rate).}

\item{nchains}{Number of Markov chains, when nchains>1, the function calculates the Gelman-Rubin convergence statistic, as modified by Brooks and Gelman (1998).}

\item{a1, a2}{Hyperparameters of kappa. Default is a1=0.1 and a2=1.}

\item{c1, c2}{Hyperparameters of tau2. Default is c1=0.1 and c2=1.}

\item{sigma2}{Variance of spike (multivariate normal distribution with a diagonal covariance matrix with small variance) representing the null effect distribution. Default is 10^-3.}

\item{snpnames}{Names of variables for the group.}

\item{genename}{Name of group.}
}
\value{
\itemize{
\item mcmcchain: The list of simulation output for all parameters.
\item Criteria: genename, snpnames, log10BF, lBFDR, theta, PPA1, PPA2 and the number of studies with nonzero signal by credible interval (CI).
\item Statistics of Trait 1 for Beta_1: Summary statistics including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
\item Statistics of Trait 2  for Beta_2: Summary statistics including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
\item Other Parameters: Summary statistics for kappa and tau2 including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
}
}
\description{
Run a Gibbs sampler for a multivariate Bayesian sparse group selection model with continuous spike prior for detecting pleiotropic effects on two traits. This function is designed for summary statistics containing estimated regression coefficients and their estimated covariance matrices.
}
\details{
Run a Gibbs sampler using a continuous spike.
}
\seealso{
  \code{\link{plot_acf_CS},\link{plot_density_CS},\link{plot_trace_CS},\link{DS},\link{HS}}
}
\examples{
mg=10
SD=diag(0.05,mg)
corr=0.5
R=matrix(corr,mg,mg)+(1-corr)*diag(mg)
Sigma=crossprod(crossprod(SD,R),SD)
sign=rbinom(mg,1,.5)
sign[sign==0]=-1
betat=rep(1,mg)
betat=betat*sign
Betah=mvtnorm::rmvnorm(2,betat,Sigma)

snpnames=1:mg
genename="simulated_data"

Betah1=Betah[1,]; Betah2=Betah[2,];
Sigmah1=Sigma; Sigmah2=Sigma;
kappa0=0.5; kappastar0=0.5; sigma20=1; mg=mg;
a1=0.1; a2=0.1; d1=0.1; d2=0.1; c1=1; c2=1; s20=1; N00=100
pvalue=2*pnorm(-abs(Betah/sqrt(diag(Sigma))))

zinit=rep(0,2)
for(j in 1:2){
  index=1:mg
  PVALUE=p.adjust(pvalue[j,])
  SIGNALS=index[PVALUE<0.05]
  modelf1=rep(0,mg)
  modelf1[SIGNALS]=1
  if(max(modelf1)==1)(zinit[j]=1)
}


res1 = CS(Betah1, Betah2,
          Sigmah1, Sigmah2,
          kappa0=0.5, tau20=1,
          zeta10=zinit[1], zeta20 =zinit[2],
          mg=mg, niter=1000, burnin=500,
          nchains=1, nthin=2, a1=.1, a2=1, c1=0.1, c2=1, sigma2=10^-3,
          snpnames, genename)

res2 = CS(Betah1, Betah2,
          Sigmah1, Sigmah2,
          kappa0=c(0.5,0.25,0.5), tau20=c(1,1.25,1.5),
          zeta10=rep(zinit[1],3), zeta20 =rep(zinit[2],3),
          mg=mg, niter=1000, burnin=500,
          nchains=3, nthin=5, a1=.1, a2=1, c1=0.1, c2=1,
          sigma2=10^-3, snpnames, genename)
}
\references{
T. Baghfalaki, P.E. Sugier, T. Truong, A.T. Pettitt, K. Mengersen and B. Liquet. (2020). Bayesian meta-analysis models to Cross Cancer Genomic Investigation of pleiotropic effects using group structure. \emph{Submitted in Statistics in Medicine}.
}
\author{
Taban Baghfalaki.
}
