% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS.R
\name{e2_Monte_Carlo_EM}
\alias{e2_Monte_Carlo_EM}
\title{Internal: e2_Monte_Carlo_EM}
\usage{
e2_Monte_Carlo_EM(
  Betah1 = Betah1,
  Betah2 = Betah2,
  Sigmah1 = Sigmah1,
  Sigmah2 = Sigmah2,
  kappa0 = kappa0,
  kappastar0 = kappastar0,
  sigma20 = sigma20,
  s20 = s20,
  mg,
  a1 = a1,
  a2 = a2,
  d1 = d1,
  d2 = d2,
  c1 = c1,
  c2 = c2,
  e2 = e2,
  snpnames,
  genename
)
}
\arguments{
\item{Betah1}{A numerical vector of length mg representing the regression coefficients for the first trait.}

\item{Betah2}{A numerical vector of length mg representing the regression coefficients for the second trait.}

\item{Sigmah1}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix Betah1.}

\item{Sigmah2}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix Betah2.}

\item{kappa0}{Initial value for kappa.}

\item{kappastar0}{Initial value for kappastar.}

\item{sigma20}{Initial value for sigma2.}

\item{s20}{Initial value for s2.}

\item{mg}{Number of variables in the group.}

\item{a1, a2}{Hyperparameters of kappa. Default is a1=0.1 and a2=0.1.}

\item{d1, d2}{Hyperparameters of sigma2. Default is d1=0.1 and d2=0.1.}

\item{c1, c2}{Hyperparameters of kappastar. Default is c1=1 and c2=1.}

\item{e2}{Initial value for doing Monte Carlo EM algorithm to estimate hyperparameter of s2.}

\item{snpnames}{Names of variables for the group.}

\item{genename}{Name of group.}
}
\description{
Internal: e2_Monte_Carlo_EM
}
