% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCPBayes.R
\docType{package}
\name{GCPBayes}
\alias{GCPBayes}
\title{GCPBayes Package}
\description{
Run a Gibbs sampler for a multivariate Bayesian sparse group selection model with Dirac, continoues and hierarchical spike prior for detecting pleiotropic effects on two traits. This package is designed for summary statistics containing estimated regression coefficients and their estimated covariance matrices.
}
\references{
T. Baghfalaki, P.E. Sugier, T. Truong, A.T. Pettitt, K. Mengersen and B. Liquet. (2020). Bayesian meta-analysis models to Cross Cancer Genomic Investigation of pleiotropic effects using group structure. Submitted in Statistics in Medicine.
}
\seealso{
  \code{\link{CS},\link{DS},\link{HS}}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, \email{t.baghfalaki@modares.ac.ir}
}
\keyword{CS,}
\keyword{DS,}
\keyword{HS,}
\keyword{plot_acf_CS,}
\keyword{plot_acf_DS,}
\keyword{plot_acf_HS,}
\keyword{plot_density_CS,}
\keyword{plot_density_DS,}
\keyword{plot_density_HS,}
\keyword{plot_trace_CS,}
\keyword{plot_trace_DS,}
\keyword{plot_trace_HS}
