% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HS.R
\name{HS}
\alias{HS}
\title{Hierarchical Spike}
\usage{
HS(
  Betah1 = Betah1,
  Betah2 = Betah2,
  Sigmah1 = Sigmah1,
  Sigmah2 = Sigmah2,
  kappa0 = kappa0,
  kappastar0 = kappastar0,
  sigma20 = sigma20,
  s20 = s20,
  mg,
  niter = 1000,
  burnin = 500,
  nthin = 2,
  nchains = 2,
  a1 = 0.1,
  a2 = 0.1,
  d1 = 0.1,
  d2 = 0.1,
  c1 = 1,
  c2 = 1,
  e2 = 1,
  snpnames,
  genename
)
}
\arguments{
\item{Betah1}{A numerical vector of length mg representing the regression coefficients for the first trait.}

\item{Betah2}{A numerical vector of length mg representing the regression coefficients for the second trait.}

\item{Sigmah1}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix Betah1.}

\item{Sigmah2}{A mg*mg positive definite covariance matrix where it is estimated covariance matrix Betah2.}

\item{kappa0}{Initial value for kappa (its dimension is equal to nchains).}

\item{kappastar0}{Initial value for kappastar (its dimension is equal to nchains).}

\item{sigma20}{Initial value for sigma2 (its dimension is equal to nchains).}

\item{s20}{Initial value for s2 (its dimension is equal to nchains).}

\item{mg}{Number of variables in the group.}

\item{niter}{Number of iterations for the Gibbs sampler.}

\item{burnin}{Number of burn-in iterations.}

\item{nthin}{The lag of the iterations used for the posterior analysis is defined (or thinning rate).}

\item{nchains}{Number of Markov chains, when nchains>1, the function calculates the Gelman-Rubin convergence statistic, as modified by Brooks and Gelman (1998).}

\item{a1, a2}{Hyperparameters of kappa. Default is a1=0.1 and a2=0.1.}

\item{d1, d2}{Hyperparameters of sigma2. Default is d1=0.1 and d2=0.1.}

\item{c1, c2}{Hyperparameters of kappastar. Default is c1=1 and c2=1.}

\item{e2}{Initial value for doing Monte Carlo EM algorithm to estimate hyperparameter of s2.}

\item{snpnames}{Names of variables for the group.}

\item{genename}{Name of group.}
}
\value{
\itemize{
\item mcmcchain: The list of simulation output for all parameters.
\item Criteria: genename, snpnames, the number of studies with nonzero signal by median and the number of studies with nonzero signal by credible interval (CI).
\item Statistics of Trait 1 for Beta_1: Summary statistics including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
\item Statistics of Trait 2  for Beta_2: Summary statistics including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
\item Other Parameters: Summary statistics for kappa, kappastar and sigma2 including Mean, SD, val2.5pc, Median,	val97.5pc and BGR (if nchains>1).
}
}
\description{
Run a Gibbs sampler for a multivariate Bayesian sparse group selection model with hierarchical spike prior for detecting pleiotropic effects on two traits. This function is designed for summary statistics containing estimated regression coefficients and their estimated covariance matrices.
}
\details{
Run a Gibbs sampler using a hierarchical spike.
}
\seealso{
  \code{\link{plot_acf_HS},\link{plot_density_HS},\link{plot_trace_HS},\link{CS},\link{DS}}
}
\examples{
mg=10
SD=diag(0.05,mg)
corr=0.5
R=matrix(corr,mg,mg)+(1-corr)*diag(mg)
Sigma=crossprod(crossprod(SD,R),SD)
sign=rbinom(mg,1,.5)
sign[sign==0]=-1
Betat=rep(1,mg)
Betat=Betat*sign
Betah=mvtnorm::rmvnorm(2,Betat,Sigma)

snpnames=1:mg
genename="simulated_data"


#######################
Betah1=Betah[1,]; Betah2=Betah[2,];
Sigmah1=Sigma; Sigmah2=Sigma;
a1=0.1; a2=0.1; d1=0.1; d2=0.1; c1=1; c2=1; e2=1; N00=100



res1 = HS(Betah1, Betah2,
          Sigmah1, Sigmah2, kappa0 = 0.5, kappastar0=0.5, sigma20 =1, s20 =1,
          mg=mg, niter=1000, burnin=500, nthin=3, nchains=1, a1=a1, a2=a2, d1=d1,
          d2=d2, c1=c1, c2=c2, e2=e2, snpnames, genename)



res2 = HS(Betah1, Betah2,
          Sigmah1, Sigmah2, kappa0 = c(0.5,0.25,0.7), kappastar0=c(0.5,0.25,0.2),
          sigma20 =c(1,1.5,2), s20 =c(1,1.5,2),
          mg=mg, niter=1000, burnin=500, nthin=3, nchains=3, a1=a1, a2=a2, d1=d1,
          d2=d2, c1=c1, c2=c2, e2=e2, snpnames, genename)

}
\references{
T. Baghfalaki, P.E. Sugier, T. Truong, A.T. Pettitt, K. Mengersen and B. Liquet. (2020). Bayesian meta-analysis models to Cross Cancer Genomic Investigation of pleiotropic effects using group structure. \emph{Submitted in Statistics in Medicine}.
}
\author{
Taban Baghfalaki.
}
